+++ src/components/Box/style/index.ts
@@ -4,4 +4,3 @@ import StyledFlex from "./flex";
 export { StyledDiv, StyledFlex };
 export { StyledCenter, StyledShadowCenter } from "./center";
 export { StyledHStack, StyledVStack } from "./stack";
-export { StyledSidebar, StyledSidebarHeader } from "./sidebar";